Use float for pressure_height.
git-svn-id: https://svn.met.no/kvoss/kvbufrd/trunk@5121 422d51e2-5941-41ac-8e65-d5b5931239be

+++ src/StationInfo.h
@@ -418,7 +418,7 @@ private:
    int            height_;
    int            heightVisability_;
    int            heightTemperature_;
-   int            heightPressure_;
+   float          heightPressure_;
    int            heightPrecip_;
    int            heightWind_;
    int            heightWindAboveSea_;
@@ -475,7 +475,7 @@ public:
    void      height( int h, bool ifUnset = true );
    int       heightVisability() const;
    int       heightTemperature()const;
-   int       heightPressure() const;
+   float     heightPressure() const;
    int       heightPrecip()const;
    int       heightWind()const;
    int       heightWindAboveSea()const;
